app/requirements.txt
fastapi
uvicorn[standard]
pydantic
openai-whisper
soundfile